+++ app/src/map/components/MapDashboard.jsx
@@ -3,28 +3,39 @@ import classnames from 'classnames';
 import PropTypes from 'prop-types';
 import ControlPanel from 'mapPanels/rightControlPanel/containers/ControlPanel';
 import Timebar from 'timebar/containers/Timebar';
-import MiniGlobe from 'mapPanels/leftControlPanel/components/MiniGlobe';
-import MobileLeftExpand from 'mapPanels/leftControlPanel/components/MobileLeftExpand';
 import ReportPanel from 'report/containers/ReportPanel';
 import MapFooter from 'siteNav/components/MapFooter';
 import LeftControlPanel from 'mapPanels/leftControlPanel/containers/LeftControlPanel';
 import Map from 'map/containers/Map';
-import mapPanelsStyles from 'styles/components/map-panels.scss';
+import MapPanelsStyles from 'styles/components/map-panels.scss';
 import mapStyles from 'styles/components/map.scss';
 
 class MapDashboard extends Component {
+  constructor() {
+    super();
+    this.state = {
+      expanded: true
+    };
+  }
+  onMapPanelsToggleExpand = () => {
+    this.setState({
+      expanded: !this.state.expanded
+    });
+  }
   render() {
     const { isEmbedded, openSupportFormModal, onExternalLink } = this.props;
     return (<div className="fullHeightContainer" >
       {!isEmbedded &&
       <div
         className={classnames(
-          mapPanelsStyles.mapPanels,
+          MapPanelsStyles.mapPanels,
           {
-            [mapPanelsStyles._noFooter]: !COMPLETE_MAP_RENDER
+            [MapPanelsStyles._noFooter]: !COMPLETE_MAP_RENDER,
+            [MapPanelsStyles._expanded]: this.state.expanded
           }
         )}
       >
+        <div className={MapPanelsStyles.expandButton} onClick={this.onMapPanelsToggleExpand} />
         <ControlPanel isEmbedded={isEmbedded} />
         <ReportPanel />
       </div >
@@ -41,15 +52,6 @@ class MapDashboard extends Component {
         <Map setAttribution={this.setAttribution} />
         <LeftControlPanel />
       </div>
-      <MobileLeftExpand isEmbedded={this.props.isEmbedded}>
-        <MiniGlobe
-          center={{ lat: this.props.latitude, lng: this.props.longitude }}
-          zoom={this.props.zoom}
-          viewportWidth={1000}
-          viewportHeight={600}
-          isEmbedded={this.props.isEmbedded}
-        />
-      </MobileLeftExpand>
       <div className={classnames(mapStyles.timebarContainer, { [mapStyles._noFooter]: !COMPLETE_MAP_RENDER })} >
         <Timebar />
       </div >